+++ tensorboard/plugins/custom_plugin/run.bat
@@ -1,2 +1,2 @@
 python setup.py develop
-tensorboard --logdir=main_plugin/logs --enable_flops --enable_system_performance
+tensorboard --logdir=main_plugin/logs
